# backend/tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kivai Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/video/mp4-to-mov/inspect": {
      "post": {
        "summary": "Inspect Mp4 To Mov",
        "operationId": "inspect_mp4_to_mov_video_mp4_to_mov_inspect_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_inspect_mp4_to_mov_video_mp4_to_mov_inspect_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/video/mp4-to-mov": {
      "post": {
        "summary": "Mp4 To Mov",
        "operationId": "mp4_to_mov_video_mp4_to_mov_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_mp4_to_mov_video_mp4_to_mov_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_inspect_mp4_to_mov_video_mp4_to_mov_inspect_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_inspect_mp4_to_mov_video_mp4_to_mov_inspect_post"
      },
...
      "Body_mp4_to_mov_video_mp4_to_mov_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "auto"
          },
          "resolution": {
            "type": "string",
            "title": "Resolution",
            "default": "original"
          },
          "fps": {
            "type": "string",
            "title": "Fps",
            "default": "original"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_mp4_to_mov_video_mp4_to_mov_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}